Refatora para utilização de js para retorno de erros. Formulários alt…

+++ app/controllers/sessions_controller.rb
@@ -2,15 +2,18 @@ class SessionsController < ApplicationController
   skip_before_action :authorized, only: :create
 
   def create
-    respond_to do |format|
-      result = AutenticateUserService.new(login: login_params).call
-      session[:user_id] = result.data.id.to_s if result.success?
-      format.html { redirect_to home_index_path, handle_message(result) }
+    result = AutenticateUserService.new(login: login_params).call
+
+    if result.success?
+      session[:user_id] = result.data.id.to_s
+      redirect home_index_path
+    else
+      notify result.error
     end
   end
 
   def destroy
-    session.delete(:user_id)
+    session.delete(:user_id) if session
     redirect_to root_path
   end
 
@@ -18,11 +21,5 @@ class SessionsController < ApplicationController
 
   def login_params
     params.require(:login).permit(:email, :password)
-  end
-
-  def handle_message result
-    if result.success?
-      { alert: result.error }
-    end
   end
 end
